refactor(wave): extract ContainerRequestHandler SPI from ContainerController
Move all container request logic (submit, findRecord, findStatus, revoke)
into a new ContainerRequestHandler base class so both the alpha controller
and the upcoming ContainersV1Controller can share the same implementation.
ContainerController now extends ContainerRequestHandler with no behaviour change.
